+++ src/font/mod.rs
@@ -22,20 +22,47 @@ pub struct Font {
     texture: Texture,
 }
 
+/// Font data and metrics.
+///
+/// Used to construct a new font
 #[derive(Debug)]
 pub struct FontData<'a> {
+    /// the height of the font
     pub height: f32,
+    /// the gap between 2 different lines
     pub line_gap: f32,
+    /// the highest point above the baseline
     pub ascender: f32,
+    /// the lowest point below the baseline (negative)
     pub descender: f32,
+    /// data for all loaded characters
     pub char_data: HashMap<char, CharData>,
+    /// grayscale bytes for creating the font atlas texture
     pub texture_data: &'a [u8],
+    /// atlas texture's width
     pub texture_width: u32,
+    /// atlas texture's height
     pub texture_height: u32,
 }
 
+/// Character data and metrics.
+///
+/// Used to construct a character
+/// in a font
 #[derive(Debug, Copy, Clone)]
-pub struct FontMetrics {
+pub struct CharData {
+    /// top-left and bottom-right corner for character's texture
+    pub uvs: Vec4,
+    /// width and height of the character's texture
+    pub bounds: Vec2,
+    /// offset of the character
+    pub bearing: Vec2,
+    /// advance between this and the next character in a string
+    pub advance: f32,
+}
+
+#[derive(Debug, Copy, Clone)]
+pub(crate) struct FontMetrics {
     pub(crate) height: f32,
     pub(crate) line_gap: f32,
     pub(crate) ascender: f32,
@@ -43,37 +70,35 @@ pub struct FontMetrics {
     pub(crate) space_width: f32,
 }
 
-#[derive(Debug, Copy, Clone)]
-pub struct CharData {
-    pub uvs: Vec4,
-    pub bounds: Vec2,
-    pub bearing: Vec2,
-    pub advance: f32,
-}
-
 impl Font {
     pub(crate) fn fira_mono(device: &Device, uniforms: &mut Uniforms) -> Result<Self> {
-        let font_data = fira_mono();
+        Self::new(device, uniforms, fira_mono())
+    }
 
+    pub(crate) fn new(
+        device: &Device,
+        uniforms: &mut Uniforms,
+        data: FontData<'_>,
+    ) -> Result<Self> {
         let texture = Texture::new(
             device,
             uniforms,
-            font_data.texture_data.to_vec(),
-            font_data.texture_width,
-            font_data.texture_height,
+            data.texture_data.to_vec(),
+            data.texture_width,
+            data.texture_height,
             Format::Gray,
             Mips::Zero,
         )?;
 
         Ok(Self {
             metrics: FontMetrics {
-                height: font_data.height,
-                line_gap: font_data.line_gap,
-                ascender: font_data.ascender,
-                descender: font_data.descender,
+                height: data.height,
+                line_gap: data.line_gap,
+                ascender: data.ascender,
+                descender: data.descender,
                 space_width: 1.0 / 3.0,
             },
-            char_data: font_data.char_data,
+            char_data: data.char_data,
             texture,
         })
     }
@@ -89,7 +114,11 @@ impl Font {
     pub(crate) fn char_data(&self, c: char) -> CharData {
         match self.char_data.get(&c) {
             Some(data) => *data,
-            None => *self.char_data.get(&'?').expect("bad default"),
+            None => *self
+                .char_data
+                .values()
+                .next()
+                .expect("font has no characters"),
         }
     }